Adding order details 2/2
