+++ src/components/OrderDetails/orderDetails.style.js
@@ -1,0 +1,6 @@
+import styled from 'styled-components';
+
+export const OrderDetailsContainer = styled.div`
+  margin: auto;
+  width: 75%;
+`;
